NRS-7 equation upgrade: rename solve rational modules
