+++ src/solve/mod.rs
@@ -8,5 +8,5 @@
 pub mod equation;
 mod complex_integer_equations;
 mod complex_float_equations;
-mod comp_rat_int_eq;
-mod comp_rat_float_eq;
+mod rational_integer_equation;
+mod rational_float_equation;
