+++ src/solve/rational_integer_equation.rs
